Renaming classes, first test LogFilter
